% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avrgTimeComparison.R
\name{avrgTimeComparison}
\alias{avrgTimeComparison}
\title{Compare two or more scenarios for averages through time}
\usage{
avrgTimeComparison(
  ...,
  upload,
  outputFolder,
  comparisonID,
  folderID,
  plotCI = TRUE
)
}
\arguments{
\item{...}{List or tables of results coming from usefun::meanValuesTime to be compared.}

\item{upload}{logical. Should the raster be updated to googledrive? Only works if the
raster exists OR is set to be written.}

\item{outputFolder}{character. Path to the folder where it should be saved.}

\item{comparisonID}{character. Name to indentify the comparison (for file name)}

\item{folderID}{character. Google folder id to upload to.
Only needs to be provided if `upload == TRUE`. Default is `NULL``}

\item{plotCI}{logical. Should the plot have confidence interval?}
}
\value{
list of significant species or scenarios with indication of increasing or decreasing
}
\description{
Compare two or more scenarios for averages through time
}
\author{
Tati Micheletti
}
