% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{binarize.to.thres}
\alias{binarize.to.thres}
\title{Binarize matrix to given threshold}
\usage{
binarize.to.thres(mat, thres)
}
\arguments{
\item{mat}{a matrix or data.frame object}

\item{thres}{a positive numerical value}
}
\description{
Simple function that checks every element of a given matrix (or data.frame)
if it surpasses the given threshold either positively or negatively and it
outputs 1 for that element otherwise 0.
}
\examples{

mat = matrix(data = -4:4, nrow = 3, ncol = 3)
binarize.to.thres(mat, thres = 0.5)
binarize.to.thres(mat, thres = 2.5)

}
