% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassPerSpeciesYearGIF.R
\name{biomassPerSpeciesYearGIF}
\alias{biomassPerSpeciesYearGIF}
\title{Creates a GIF of biomass change}
\usage{
biomassPerSpeciesYearGIF(dataPath, years = NULL, uploadTo)
}
\arguments{
\item{dataPath}{character. Path to data}

\item{years}{numeric. Years available/intended to be used for the giphy}

\item{uploadTo}{character. Google drive folder id. (i.e. "1ZqPVs33HxnnmjLUW94i7AuwAS-nloPGH")}
}
\value{
list of plots
}
\description{
Creates a GIF of biomass change
}
\author{
Tati Micheletti
}
