% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassYearGIF.R
\name{biomassYearGIF}
\alias{biomassYearGIF}
\title{Creates a GIF of total biomass change}
\usage{
biomassYearGIF(dataPath, years = NULL, uploadTo)
}
\arguments{
\item{dataPath}{character. Path to data}

\item{years}{numeric. Years available/intended to be used for the giphy}

\item{uploadTo}{character. Google drive folder id. (i.e. "1ZqPVs33HxnnmjLUW94i7AuwAS-nloPGH")}
}
\value{
RasterLtack
}
\description{
Creates a GIF of total biomass change
}
\author{
Tati Micheletti
}
