% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapPercentChanges.R
\name{bootstrapPercentChanges}
\alias{bootstrapPercentChanges}
\title{Bootstraps rasters for testing significancy on comparable rasters of different species or scenarios}
\usage{
bootstrapPercentChanges(
  dataPath,
  years = c(2001, 2100),
  sampleSize = "auto",
  n = 100,
  shp = NULL,
  species = NULL,
  useFuture = FALSE
)
}
\arguments{
\item{dataPath}{character. Path to raster data.}

\item{years}{numeric. Years to compare. Currently this function only compares 2 years.}

\item{sampleSize}{numeric or "auto" (default). What is the sample size (i.e. number of pixels)
we want to use on the bootstrapping? If "auto" it calculates internally
Cohen's D And Hedges G Effect Size.}

\item{n}{numeric. Default is 100. How many iterations (random selection of `sampleSize` pixels) should be done?}

\item{shp}{character or shapefile. If you wish to calculate these metrics for separate areas.
Needs to match the rasters. Default is NULL (i.e. the whole raster is only one area)}

\item{species}{character. Default is NULL. Which species should this function be ran ?}

\item{useFuture}{logical. Should use future to parallelize? Requires `future` and `future_apply`` packages}
}
\value{
list of significant species or scenarios with indication of increasing or decreasing
}
\description{
Bootstraps rasters for testing significancy on comparable rasters of different species or scenarios
}
\author{
Tati Micheletti
}
