% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindFromList.R
\name{cbindFromList}
\alias{cbindFromList}
\title{cbindFromList: column binds a list of data frames or data.tables, returning a merged table with unique columns}
\usage{
cbindFromList(lst)
}
\arguments{
\item{lst}{List of data.tables or data.frames to column bind}
}
\value{
Returns a merged data.table
}
\description{
cbindFromList: column binds a list of data frames or data.tables, returning a merged table with unique columns
}
\author{
Tati Micheletti
}
