% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeTraits.R
\name{changeTraits}
\alias{changeTraits}
\title{changeTraits provides an easy way of modifying species trait's table in LandR by 
multiplication of original traits by a specified factor}
\usage{
changeTraits(speciesTable, param, facMult, species)
}
\arguments{
\item{speciesTable}{speciesTable data.table from LandR. Contains different traits in columns and 
a row for each species.}

\item{param}{Character vector of the trait to be modified.}

\item{facMult}{numeric. Factor by which the specified trait should be multiplied.}

\item{species}{character string of the Species for which the trait should be modified}
}
\value{
Returns a data.table updated with the new parameters.
}
\description{
changeTraits provides an easy way of modifying species trait's table in LandR by 
multiplication of original traits by a specified factor
}
\author{
Tati Micheletti
}
