% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackagesVersions.R
\name{checkPackagesVersions}
\alias{checkPackagesVersions}
\title{All packages' versioning}
\usage{
checkPackagesVersions(simList = NULL, filePath = NULL)
}
\arguments{
\item{simList}{simList object resulting from a SpaDES call. Default is `NULL`.}

\item{filePath}{Character string. If passed, the table is written to the specified file path,
as an `RDS` file. Default is `NULL`.}
}
\description{
Returns a data.table of versions of all packages attached to your current R session, or used in
a simulation - including git commit if a package was installed from github and similars.
}
\details{
If you don't provide it, the function will return the information regarding your
         current R session's informartion. When running a simulation, it is automatically
         be created by the `spades()` and attached as the invisible object `.packagesVersions`.
}
\author{
Tati Micheletti
}
