% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{colors.100}
\alias{colors.100}
\title{100 distinct colors}
\format{
An object of class \code{character} of length 100.
}
\usage{
colors.100
}
\description{
100 as-much-as-possible distinct colors!
}
\keyword{datasets}
