% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{createModels}
\alias{createModels}
\title{Designed as a helper to create caribou models based on 2011 ECCC report on population demographics and RSF.
This function uses \code{usefun::createEquation()} internally}
\usage{
createModels(caribouCoefTable = NULL, modelsToUse = "M3")
}
\arguments{
\item{caribouCoefTable}{data.table of the coefficient values for all models available. If not provided,
will use as default the one from ECCC 2011 for population demographics.}

\item{modelsToUse}{Character string of the model to be used. Currently only "M3" and "M7" (demograohics)
and "TaigaPlains" (RSF) have been implemented. Default is "M3".}
}
\value{
names list of the quote of the model that can be parsed to generate the simulated data bootstrapped
        using the model coefficients.
}
\description{
Designed as a helper to create caribou models based on 2011 ECCC report on population demographics and RSF.
This function uses \code{usefun::createEquation()} internally
}
\author{
Tati Micheletti
}
