% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModObject.R
\name{createModObject}
\alias{createModObject}
\title{This function checks a sim list for a specific object at a specific time, or
searches for it in an input folder (i.e. saved outputs). This function
simulates the existence of a sim list with specific objects in time t.}
\usage{
createModObject(data, sim = NULL, pathInput, currentTime, fun = readRDS)
}
\arguments{
\item{data}{character string of the object or file name to be searched for}

\item{sim}{simList. Default is `NULL`.}

\item{pathInput}{path to the directory containing the files to be searched for.}

\item{currentTime}{Numeric. Current time to be used to serach for the object.}

\item{fun}{function to be used to recover the file. Default is `readRDS`}
}
\value{
A list of the objects that were read from disk. Currently the
            function only works with `.tif` (i.e. raster) and `.rds` objects.
}
\description{
This function checks a sim list for a specific object at a specific time, or
searches for it in an input folder (i.e. saved outputs). This function
simulates the existence of a sim list with specific objects in time t.
}
\author{
Tati Micheletti
}
