% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShrubHerbLayers.R
\name{createShrubHerbLayers}
\alias{createShrubHerbLayers}
\title{createShrubHerbLayers creates the necessary shrub or herb layers for predictive caribou RSF module. 
It is a function designed to work inside a SpaDES module.}
\usage{
createShrubHerbLayers(
  landCoverECCC,
  reclassLCC05,
  layerName,
  includeCrops = FALSE
)
}
\arguments{
\item{landCoverECCC}{RasterLayer LCC2005 reclassified to ECCC land cover.}

\item{reclassLCC05}{List with reclassification for LCC05 values
(i.e. LCC05 classes that should be classified as shrub or herbs)}

\item{layerName}{character. Name of the shrub layer in the model.}

\item{includeCrops}{Logical. Should this layer include crops?}
}
\value{
RasterLayer of herbs or shrubs
}
\description{
createShrubHerbLayers creates the necessary shrub or herb layers for predictive caribou RSF module. 
It is a function designed to work inside a SpaDES module.
}
\author{
Tati Micheletti
}
