% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDisturbanceFast.R
\name{extractDisturbanceFast}
\alias{extractDisturbanceFast}
\title{extractDisturbanceFast extracts disturbance of anthropogenic and fire}
\usage{
extractDisturbanceFast(
  ageMap,
  caribouShapefile,
  recoveryTime = 40,
  anthropogenicLayer = NULL,
  waterRaster,
  rasterToMatch
)
}
\arguments{
\item{ageMap}{RasterLayer. Map with forest age.}

\item{caribouShapefile}{Shapefile with polygons for which we want to calculate lambda for the caribou demographic model.}

\item{recoveryTime}{numeric. Recovery time in years that the forest needs to support Caribou. Default = 40.}

\item{anthropogenicLayer}{Anthropogenic disturbance (raster) layer. Currently, road density layer used for both RSF and demographic models.}

\item{waterRaster}{Raster layer indicating water bodies.}

\item{rasterToMatch}{RasterLayer template for these layers to match.}
}
\value{
A list of the anthropogenic and fire disturbances as percent (0-100)
}
\description{
extractDisturbanceFast extracts disturbance of anthropogenic and fire
}
\author{
Tati Micheletti
}
