% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{get.parent.dir}
\alias{get.parent.dir}
\title{Retrieve the parent directory}
\usage{
get.parent.dir(pathStr)
}
\arguments{
\item{pathStr}{string. The name of the directory, can be a full path filename.}
}
\value{
a string representing the parent directory. When a non-file path is
used as input (or something along those lines :) then it returns the root
("/") directory.
}
\description{
Use this function to retrieve the parent directory from a string representing
the full path of a file or a directory.
}
\examples{
get.parent.dir("/home/john")
get.parent.dir("/home/john/a.txt")
get.parent.dir("/home")

}
