% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_ternary_class_id}
\alias{get_ternary_class_id}
\title{Get ternary class id}
\usage{
get_ternary_class_id(value, threshold)
}
\arguments{
\item{value}{numeric}

\item{threshold}{numeric}
}
\value{
an integer. There are 3 cases:
\itemize{
\item \eqn{1}: when \eqn{value > threshold}
\item \eqn{-1}: when \eqn{value < -threshold}
\item \eqn{0}: otherwise
}
}
\description{
Helper function that checks if a \emph{value} surpasses the given
\emph{threshold} either positively, negatively or not at all and returns
a value indicating in which class (i.e. interval) it belongs.
}
