% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayers.R
\name{getLayers}
\alias{getLayers}
\title{getLayers gets the necessary layers for predictive modules such as caribouPopGrowth and caribouRSF.
It is a function designed to work inside a SpaDES module.}
\usage{
getLayers(
  currentTime,
  cohortData,
  pixelGroupMap,
  startTime,
  endTime,
  recoveryTime = 40,
  listSACaribou,
  anthropogenicLayer,
  roadDensity,
  waterRaster,
  isRSF = FALSE,
  decidousSp = NULL,
  oldBurnTime = NULL,
  elevation = NULL,
  vrug = NULL,
  LCC05 = NULL,
  reclassLCC05 = NULL,
  rasterToMatch = NULL
)
}
\arguments{
\item{currentTime}{numeric. Current time being used (i.e. \code{time(sim)}).}

\item{cohortData}{data.table. Output from LandR_Biomass module.}

\item{pixelGroupMap}{raster to identify the cohortData.}

\item{startTime}{numeric. startTime of the simulation. Needed to verify and potentially adjust relative simulation times.}

\item{endTime}{numeric. endTime of the simulation. Needed to verify and potentially adjust relative simulation times.}

\item{recoveryTime}{numeric. Recovery time in years that the forest needs to support Caribou. Default = 40.}

\item{listSACaribou}{list of shapefiles with polygons for which we want to calculate lambda for the caribou demographic model.}

\item{anthropogenicLayer}{Anthropogenic disturbance (raster) layer. Currently, 500m buffered anthropogenic disturbance for demographic models.}

\item{roadDensity}{Anthropogenic disturbance (raster) layer. Currently, road density layer used for RSF models.}

\item{waterRaster}{Raster layer indicating water bodies.}

\item{isRSF}{logical. Identify if it should get the layers for the RSF or demographic model.}

\item{decidousSp}{binary raster layer indicating if the dominant biomass in a pixel belongs to a deciduous species.}

\item{oldBurnTime}{numeric. Definition of the initial interval considered to be old burn. The end of this time is 20 years later (i.e. 40-60 years).}

\item{elevation}{RasterLayer of elevation}

\item{vrug}{RasterLayer of ruggeness}

\item{LCC05}{RasterLayer of landcover classes 2005}

\item{reclassLCC05}{List with reclassification for LCC05 values (i.e. LCC05 classes that should be classified as shrub or herbs)}

\item{rasterToMatch}{RasterLayer template for these layers to match.}
}
\value{

}
\description{
getLayers gets the necessary layers for predictive modules such as caribouPopGrowth and caribouRSF.
It is a function designed to work inside a SpaDES module.
}
\author{
Tati Micheletti
}
