% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepMulti.R
\name{grepMulti2}
\alias{grepMulti2}
\title{grepMulti works similarly to \code{grepl}, but for multiple patterns and returning the object.}
\usage{
grepMulti2(x, patterns, unwanted = NULL)
}
\arguments{
\item{x}{object where to look for patterns.}

\item{patterns}{Character vector of patterns to look for objects.}

\item{unwanted}{Character vector of patterns to exclude from search.}
}
\value{
The objects with specified patterns combined
}
\description{
grepMulti works similarly to \code{grepl}, but for multiple patterns and returning the object.
}
\author{
Tati Micheletti
}
