\name{individual.R2}
\alias{individual.R2}
\title{Individual \eqn{R^2} for each predictor of \code{lm()} object}
\description{ Finds how much each predictor contributes to the total variance explained (\eqn{R^2}) in any given linear model save as an \code{lm()} object}
\usage{individual.R2(x)}
\arguments{
  \item{x}{an \emph{lm} object generated with the \code{lm()} base R function}
          }
\value{ Returns a dataframe object}
