% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{is.empty}
\alias{is.empty}
\title{Is object empty?}
\usage{
is.empty(obj)
}
\arguments{
\item{obj}{a general object}
}
\value{
a logical specifying if the object is NULL or not.
}
\description{
A function to test whether an object is \strong{NULL, empty} or something
similar :) It checks the length of the object, so it has different behaviour
than \code{\link{is.null}}.
}
\examples{
# TRUE
is.empty(NULL)
is.empty(c())

# FALSE
is.empty("")
is.empty(NA)
is.empty(NaN)

}
