% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_color_bar_plot}
\alias{make_color_bar_plot}
\title{Make a color bar plot}
\usage{
make_color_bar_plot(color.vector, number.vector, title,
  x.axis.label = "")
}
\arguments{
\item{color.vector}{vector of color values}

\item{number.vector}{vector of numeric values (same length with
\code{color.vector})}

\item{title}{string. The title of the barplot}

\item{x.axis.label}{string. The x-axis label. Default value: empty string}
}
\description{
Use this function when you want to visualize some numbers and their
respective color values. Note that more than 42 colors won't be nice
to see (too thin bars)!
}
\examples{
color.vector = rainbow(10)
number.vector = 1:10
title = "First 10 rainbow() colors"
make_color_bar_plot(color.vector, number.vector, title)

}
