% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{mat.equal}
\alias{mat.equal}
\title{Matrix equality}
\usage{
mat.equal(x, y)
}
\arguments{
\item{x, y}{matrices}
}
\value{
a logical specifying if the two matrices are equal or not.
}
\description{
Check if two matrices are equal. Equality is defined by both of them being
matrices in the first place, having the same dimensions as well as the same
elements.
}
