% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesAvailable.R
\name{modulesAvailable}
\alias{modulesAvailable}
\title{Function to show the modules that have stickers available}
\usage{
modulesAvailable()
}
\value{
This function returns a vector of all modules that have stickers available
}
\description{
Function to show the modules that have stickers available
}
\author{
Tati Micheletti
}
