% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleSticker.R
\name{moduleSticker}
\alias{moduleSticker}
\title{Function to create stikers for SpaDES modules}
\usage{
moduleSticker(
  moduleTable = NULL,
  moduleName,
  directory = getwd(),
  useCache = NULL,
  savedSticker = NULL,
  ...
)
}
\arguments{
\item{moduleTable}{Table with the information to generate the sticker (i.e. figure name, height, colors, etc)}

\item{moduleName}{charcter string of the module's name.}

\item{directory}{charachter string of the directory where the sticker should be saved}

\item{useCache}{logical. Should the process of downloading and loading the table be cached?}

\item{savedSticker}{Character string. Name of the file to e saved. If not provided,
saves with the name of the figure. Useful if more than one stickers
have the same figure}

\item{...}{Other parameters for ggplot2 (i.e. fontface, lineheight)}
}
\value{
This function returns the location where the sticker is available
}
\description{
Function to create stikers for SpaDES modules
}
\author{
Tati Micheletti
}
