% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAndBringOn.R
\name{nameAndBringOn}
\alias{nameAndBringOn}
\title{nameAndBringOn name's a raster, postProcess to RTM and brings it to memory.}
\usage{
nameAndBringOn(ras, name, RTM = NULL)
}
\arguments{
\item{ras}{RasterLayer.}

\item{name}{character. Name of the raster layer.}

\item{RTM}{RasterLayer template. If a RTM is passed, the function masks the raster to it, converting 
non-NA inside it to 0}
}
\value{
RasterLtack
}
\description{
nameAndBringOn name's a raster, postProcess to RTM and brings it to memory.
}
\author{
Tati Micheletti
}
