% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{normalize_vector}
\alias{normalize_vector}
\title{Normalize vector}
\usage{
normalize_vector(vec, range = c(0, 1))
}
\arguments{
\item{vec}{vector with at least two different elements}

\item{range}{vector of two elements specifying the desired normalized range.
Default value is c(0,1)}
}
\value{
the normalized vector
}
\description{
Normalize a vector of numeric values in a specified range.
}
\examples{
vec = 1:10
normalize_vector(vec)
normalize_vector(vec, range = c(-1,1))

}
