% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{outersect}
\alias{outersect}
\title{Outersect}
\usage{
outersect(x, y)
}
\arguments{
\item{x, y}{vectors}
}
\value{
a vector of the non-common elements of x and y.
}
\description{
Performs set \emph{outersection} on two vectors. The opposite operation from
\code{intersect}!
}
\examples{
x = 1:10
y = 2:11

# c(1,11)
outersect(x,y)

}
\seealso{
\code{\link{intersect}}
}
