% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_string_to_file}
\alias{plot_string_to_file}
\title{Plot string to output format}
\usage{
plot_string_to_file(
  file,
  file.format = c("pdf", "png", "svg", "tiff"),
  plot.string
)
}
\arguments{
\item{file}{string. The name of the file, can be a full path.}

\item{file.format}{string. The output file format. Can be one of these:
pdf, svg, png or tiff.}

\item{plot.string}{string. The plot string expression.}
}
\description{
Execute a plot string expression and output the result to the specified
file format.
}
\examples{

x = 1:10
y = 1:10
plot_string_to_file(paste0(tempdir(), "/my_file.pdf"), "pdf", "plot(x,y)")

}
