% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBurnSummary.R
\name{plotBurnSummary}
\alias{plotBurnSummary}
\title{Plots summary of burns}
\usage{
plotBurnSummary(
  dataPath,
  typeSim,
  lastYear,
  theObject = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dataPath}{character. Path to data}

\item{typeSim}{character. Which simulation is it? i.e. 'LandR_SCFM' | 'LandR.CS_fS'}

\item{overwrite}{logical. Default FALSE}

\item{years}{numeric. Years available/intended to be used for the giphy}

\item{saveRAS}{logical. Save the raster for posterior use?}
}
\value{
plot
}
\description{
Plots summary of burns
}
\author{
Tati Micheletti
}
