% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLeadingVegetationType.R,
%   R/plotVegetationBiomass.R
\name{plotLeadingVegetationType}
\alias{plotLeadingVegetationType}
\alias{plotVegetationBiomass}
\title{Plots Leading Vegetation Type using cohortData and pixelGroupMap}
\usage{
plotLeadingVegetationType(years = c(2011, 2100), dataPath, typeSim,
  colNA = "grey85", saveRAS = TRUE)

plotVegetationBiomass(years = c(2001, 2100), dataPath, typeSim,
  colNA = "grey85", saveRAS = TRUE)
}
\arguments{
\item{years}{numeric. Years available/intended to be used for the giphy}

\item{dataPath}{character. Path to data}

\item{typeSim}{character. Which simulation is it? i.e. 'LandR_SCFM' | 'LandR.CS_fS'}

\item{saveRAS}{logical. Save the raster for posterior use?}

\item{years}{numeric. Years available/intended to be used for the giphy}

\item{dataPath}{character. Path to data}

\item{typeSim}{character. Which simulation is it? i.e. 'LandR_SCFM' | 'LandR.CS_fS'}

\item{saveRAS}{logical. Save the raster for posterior use?}
}
\value{
plot

plot
}
\description{
Plots Leading Vegetation Type using cohortData and pixelGroupMap

Plots Leading Vegetation Type using cohortData and pixelGroupMap
}
\author{
Tati Micheletti

Tati Micheletti
}
