% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputsLayers_DUCKS.R
\name{prepInputsLayers_DUCKS}
\alias{prepInputsLayers_DUCKS}
\title{Prepare DUCKS layer}
\usage{
prepInputsLayers_DUCKS(
  destinationPath,
  lccLayer = "2005",
  url = NULL,
  archive = NULL,
  targetFile = NULL,
  studyArea = NULL,
  rasterToMatch = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{destinationPath}{Path where to save the downloaded file.}

\item{lccLayer}{Which year should be used as a base for the vegetation layer? Default to 2005}

\item{url}{The url from where the layer should be downloaded from. if \code{NULL}, the default is \code{}}

\item{studyArea}{Study area for which the layer should be cropped to}

\item{rasterToMatch}{If output rasters should}
}
\value{
RasterLayer
}
\description{
This function is intendend to prepare the DUCKS Unlimited Hybrid Wetland
             v. 2.1 layer to be used for different purposes. The output is a RasterLayer
             cropped and reprojected to the \code{studyArea}, as well as resampled to the
             \code{rasterToMatch} if any of these are provided.
}
