% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{pretty_print_bold_string}
\alias{pretty_print_bold_string}
\title{Pretty print a bold string}
\usage{
pretty_print_bold_string(string, with.gt = TRUE, html.output = TRUE)
}
\arguments{
\item{string}{a string}

\item{with.gt}{logical. Determines if the ">" sign will be appended for nice
printing in an R notebook. (use with the chuck option \emph{results = 'asis'}). Default value: TRUE.}

\item{html.output}{logical. If TRUE, it encapsulates the string with the bold
tags for an HTML document. Default value: TRUE.}
}
\description{
Prints a bold string only when \code{html.output} is enabled. Otherwise, it prints
a normal string. The the ">" sign can be appended if nice output in an R
notebook is desired.
}
\seealso{
\code{\link{pretty_print_string}}
}
