% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{pretty_print_name_and_value}
\alias{pretty_print_name_and_value}
\title{Pretty print a name and value}
\usage{
pretty_print_name_and_value(name, value, with.gt = FALSE, with.comma = TRUE)
}
\arguments{
\item{name}{string}

\item{value}{string}

\item{with.gt}{logical. Determines if the ">" sign will be appended for nice
printing in an R notebook (use with the chuck option \emph{results = 'asis'}).
Default value: FALSE.}

\item{with.comma}{logical. Determines if the comma (,) character will be
appended to the end of the output. Default value: TRUE.}
}
\description{
Pretty print a name and value
}
\examples{
pretty_print_name_and_value("aName", "aValue", with.gt = TRUE)
pretty_print_name_and_value("aName", "aValue", with.comma = FALSE)
}
