% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{pretty_print_string}
\alias{pretty_print_string}
\title{Pretty print a string}
\usage{
pretty_print_string(string, with.gt = TRUE)
}
\arguments{
\item{string}{a string}

\item{with.gt}{logical. Determines if the ">" sign will be appended for nice
printing in an R notebook. Default value: TRUE.}
}
\description{
Nice printing of a string in an R notebook (default behaviour). Otherwise,
it prints the string to the standard R output.
}
\seealso{
\code{\link{cat}}
}
