% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{pretty_print_vector_names}
\alias{pretty_print_vector_names}
\title{Pretty printing of a vector's names attribute}
\usage{
pretty_print_vector_names(vec, vector.names.str = "nodes", sep = ", ",
  with.gt = TRUE)
}
\arguments{
\item{vec}{vector}

\item{vector.names.str}{string. It tell us what are the names of the
vector (use plural form) in order to fill the print message. Default value:
"nodes".}

\item{sep}{string. The separator character to use to distinguish between
the names values. Default value: ", ".}

\item{with.gt}{logical. Determines if the ">" sign will be appended for nice
printing in an R notebook (use with the chuck option \emph{results = 'asis'}).
Default value: TRUE.}
}
\description{
Pretty printing of a vector's names attribute
}
\seealso{
\code{\link{pretty_print_string}}
}
