% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{pretty_print_vector_values}
\alias{pretty_print_vector_values}
\title{Pretty printing of a vector's values}
\usage{
pretty_print_vector_values(vec, vector.values.str = "nodes",
  sep = ", ", with.gt = TRUE)
}
\arguments{
\item{vec}{vector}

\item{vector.values.str}{string. It tell us what are the values of the
vector (use plural form) in order to fill the print message. Default value:
"nodes".}

\item{sep}{string. The separator character to use to distinguish between
the vector values. Default value: ", ".}

\item{with.gt}{logical. Determines if the ">" sign will be appended for nice
printing in an R notebook. Default value: TRUE.}
}
\description{
Pretty printing of a vector's values
}
\seealso{
\code{\link{pretty_print_string}}
}
