% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_empty_line}
\alias{print_empty_line}
\title{Print an empty line}
\usage{
print_empty_line(html.output = FALSE)
}
\arguments{
\item{html.output}{logical. If TRUE, it outputs an empty line for an HTML
document, else an empty line for the standard R output. Default value:
FALSE.}
}
\description{
Print an empty line
}
\seealso{
\code{\link{cat}}
}
