% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provinceBCRStudyArea.R
\name{provinceBCRStudyArea}
\alias{provinceBCRStudyArea}
\title{Preparing study area based on BCR and Canadian or American provinces or states.}
\usage{
provinceBCRStudyArea(bcr = NULL, province = NULL, country, ...)
}
\arguments{
\item{bcr}{Numeric. Bird Conservation Region in North America that you want to crop for.
Default is `NULL`. If `NULL``, it returns only the shapefile of the province.}

\item{province}{Character string. Province or territory. Default is `NULL`. If null, returns
the map of ??.}

\item{country}{Character string. 3 letter ISO for a specific country. The complete list of
countries can be seen by calling `raster::getData('ISO3')}

\item{...}{Arguments to be passed to `prepInputs` or `Cache` (i.e. targetFile,
cacheId, destinationPath, overwrite, etc.).}
}
\description{
Downloads, reprojects, crops and masks to speficic areas in canada such as:
BCR6, random areas, provinces and territories, or any of the last in the
BCR6.
}
\details{
If you provide a bcr that is outside of a province area, the object returned is `NULL`.
}
\author{
Tati Micheletti
}
