% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{prune_rows_from_df}
\alias{prune_rows_from_df}
\title{Prune single-value rows from a data frame}
\usage{
prune_rows_from_df(df, value)
}
\arguments{
\item{df}{\code{data.frame}}

\item{value}{an integer value}
}
\value{
the row-pruned \code{data.frame}
}
\description{
Given a \code{data.frame} and an integer value, it checks whether there is a
row vector whose values match the given one. If so, it prunes that
single-valued row from the \code{data.frame}
}
\examples{
df = data.frame(c(0,0,0), c(0,1,0), c(1,0,0))
prune_rows_from_df(df, value = 0)

}
