% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSFplot.R
\name{RSFplot}
\alias{RSFplot}
\title{ECCC's RSF plot style}
\usage{
RSFplot(
  ras,
  upload = FALSE,
  writeReclasRas = FALSE,
  outputFolder = tempdir(),
  rasName,
  folderID = NULL
)
}
\arguments{
\item{ras}{RasterLayer. Layer to generate 10 binned green to red (caribou RSF) map.}

\item{upload}{logical. Should the raster be updated to googledrive? Only works if the
raster exists OR is set to be written.}

\item{writeReclasRas}{logical. Default is `FALSE`. Should the raster be written to disk?}

\item{outputFolder}{character. Path to the folder where it should be saved.}

\item{rasName}{character. Name of the raster to be saved.}

\item{folderID}{character. Google folder id to upload to.
Only needs to be provided if `upload == TRUE`. Default is `NULL``}
}
\value{
Plots that have similat binning system to ECCC's 2011 Caribou RSF
}
\description{
ECCC's RSF plot style
}
\author{
Tati Micheletti
}
