% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save.df.to.file}
\alias{save.df.to.file}
\title{Save data frame to a specified file}
\usage{
save.df.to.file(df, file)
}
\arguments{
\item{df}{data.frame}

\item{file}{string. The name of the file, can be a full path.}
}
\description{
Function for saving a \code{data.frame} to a specified file.
}
