% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save.matrix.to.file}
\alias{save.matrix.to.file}
\title{Save matrix to a specified file}
\usage{
save.matrix.to.file(mat, file)
}
\arguments{
\item{mat}{matrix}

\item{file}{string. The name of the file, can be a full path.}
}
\description{
Function for saving a \code{matrix} to a specified file. Uses the
\code{\link{save.df.to.file}} function.
}
