% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save.vector.to.file}
\alias{save.vector.to.file}
\title{Save vector to a specified file}
\usage{
save.vector.to.file(vector, file, with.row.names = FALSE)
}
\arguments{
\item{vector}{vector}

\item{file}{string. The name of the file, can be a full path.}

\item{with.row.names}{logical. If TRUE, then the \code{names(vector)} will be
included in the output file. Default value: FALSE.}
}
\description{
Function for saving a \code{vector} with or without its row names to a
specified file.
}
