% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_mat_to_file}
\alias{save_mat_to_file}
\title{Save matrix to a specified file}
\usage{
save_mat_to_file(mat, file)
}
\arguments{
\item{mat}{matrix}

\item{file}{string. The name of the file, can be a full path.}
}
\description{
Function for saving a \code{matrix} to a specified file. Uses the
\code{\link{save_df_to_file}} function.
}
