% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTempFolder.R
\name{setTempFolder}
\alias{setTempFolder}
\title{This functions sets a temporary folder for raster and other packages inside the user's cache folder.
Make sure you understand the consequences of using it.}
\usage{
setTempFolder(paths, setTmpFolder, usr)
}
\arguments{
\item{paths}{paths set with \code{SpaDES.core::setPaths()}.}

\item{setTmpFolder}{logical. Should the function reset your temp folder to a new location?}

\item{name}{caracter. User name for a final verification. If you are sure you want to set
your temporary location, use "Tati".}
}
\value{
invisible
}
\description{
This functions sets a temporary folder for raster and other packages inside the user's cache folder.
Make sure you understand the consequences of using it.
}
\author{
Tati Micheletti
}
