% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{specify.decimal}
\alias{specify.decimal}
\title{Specify decimal}
\usage{
specify.decimal(number, digits.to.keep)
}
\arguments{
\item{number}{numeric}

\item{digits.to.keep}{numeric. Refers to the digits to keep after decimal
point '.'}
}
\value{
the pruned number in string format
}
\description{
Use this function to transform a given decimal number to the desired
precision by choosing the number of digits after the decimal point.
}
\examples{
# 0.123
specify.decimal(0.1233213, 3)

}
