% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalBiomassPerSpecies.R
\name{totalBiomassPerSpecies}
\alias{totalBiomassPerSpecies}
\title{Plots biomass per species: proportional or absolute, and total or just overstory}
\usage{
totalBiomassPerSpecies(dataPath, typeSim, proportional = FALSE,
  columnsType = FALSE, overstory = FALSE, overwrite = FALSE,
  maxVal = 2e+10)
}
\arguments{
\item{dataPath}{character. Path to data}

\item{typeSim}{character. Which typeSimation is it? i.e. 'LandR_SCFM' | 'LandR.CS_fS'}

\item{proportional}{logical. Should the plot be of the proportional biomass?}

\item{columnsType}{logical. Should the plot be continuous (lines) or columns?}

\item{overstory}{logical. Should the plot be of the overstory biomass?}

\item{overwrite}{logical.}

\item{maxVal}{numeric. Max value for y axis. Passing this ensures that both overstory
and all biomass plots are comparable Default to 1e10.}
}
\value{
plot
}
\description{
Plots biomass per species: proportional or absolute, and total or just overstory
}
\author{
Tati Micheletti
}
