% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrBoth.R
\name{substrBoth}
\alias{substrBoth}
\title{substrBoth get a sub-string based on the number of characters and the side to start}
\usage{
substrBoth(strng, howManyCharacters, fromEnd = TRUE)
}
\arguments{
\item{strng}{String from which to grab a subset}

\item{howManyCharacters}{numeric. How many characters should be returned in the sub-string?}

\item{fromEnd}{logical. Default is TRUE. Should te subset start in the end of the string?}
}
\value{
character string of the subset.
}
\description{
substrBoth get a sub-string based on the number of characters and the side to start
}
\author{
Tati Micheletti
}
